Should use long variable to return a long!
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/io/trunk@1470625 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/io/input/CharSequenceInputStream.java
@@ -178,7 +178,7 @@ public class CharSequenceInputStream extends InputStream {
 
     @Override
     public long skip(long n) throws IOException {
-        int skipped = 0;
+        long skipped = 0;
         while (n > 0 && this.cbuf.hasRemaining()) {
             this.cbuf.get();
             n--;
